Refactor app.py and LexicalAnalyzer.py
I refactored the Flask app.py to remove unnecessary lines of code and
make the codebase easier to maintain.

I also refactored the lexical  analyzer to make it easier to maintain
and be traversed similar to the transition diagram. The lexer is not yet
finished as it needs to handle errors correctly which will be
implemented next.