Declutter app.py and refactor LexicalAnalyzer.py
I decluttered the lexical analyzer part of the app.py and remove
unnecessary debugging prints.

I refactored the fallback logic of identifiers into their own function
and integrate the logic of the reset position function. This makes it
easy to maintain and adheres to DRY principle. I also reverted the
structure of the inline if-else back to nested if-else due to the length
of code per line, which makes it hard for scanning.