Add lexical errors for reserved symbols
I added an invalid character lexical error for reserved symbols so that
when the expected character or delimiter is not found, the user is
alerted to the lexical error.

I also removed the original state 300 error since it is not used
anymore. Similarly, the old IllegalCharError class is removed since it
is not used anymore.